requirements.txt
asgiref==3.5.2
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
pymongo==3.12.1
pyspark==3.1.2
pytz==2022.1
sqlparse==0.2.4
gunicorn==20.1.0
whitenoise==6.1.0